+++ lib/braintree/test_subscription.rb
@@ -1,0 +1,7 @@
+module Braintree
+  class TestSubscription < Subscription
+    def self.recurring(subscription_id, date)
+      Configuration.gateway.testing.recurring(subscription_id, date)
+    end
+  end
+end