+++ lib/braintree/testing_gateway.rb
@@ -36,6 +36,20 @@ module Braintree
       @transaction_gateway._handle_transaction_response(response)
     end
 
+    def recurring(subscription_id, date)
+      check_environment
+
+      response = @config.http.put("#{@config.base_merchant_path}/subscriptions/#{subscription_id}/recurring?date=#{date}")
+      result = response[:response]
+      if result[:success]
+        SuccessfulResult.new(:success => result[:success])
+      elsif result[:error]
+        ErrorResult.new(@gateway, {errors: {error: result[:error]}})
+      end
+    rescue NotFoundError
+      raise NotFoundError, "subscription with id #{subscription_id.inspect} not found"
+    end
+
     def check_environment
       raise TestOperationPerformedInProduction if @config.environment == :production
     end